c'est pas les uint ou des pulse < 3us (vérifié à l'oscilloscope)

+++ include/capactiveSensor.h
@@ -1,6 +1,8 @@
 #ifndef CAPACITIVESENSOR_H
 #define CAPACITIVESENSOR_H
 
+#include "stdint.h"
+
 #define pulsePin 17
 #define oscilatorPin 13
 #define capacityDiffMeanFactor 3
@@ -11,8 +13,8 @@ void stopCapacitiveSensor();
 void startCapacitiveSensor();
 void resetCapacitiveSensorValue();
 
-unsigned long getCapacityDiff();
-unsigned long getCapacitiveSensorValue();
-unsigned long getCapacityDiffMean();
+int32_t getCapacityDiff();
+int32_t getCapacitiveSensorValue();
+int32_t getCapacityDiffMean();
 
 #endif
